pyproject.toml
[project]
name = "guardraills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "guardrails-ai>=0.5.15",
    "ollama>=0.4.2",
]
